Move PathFeatureExpectationsSpec to PathExpectationsSpec
