--- apis/fluent-en_GB/atrium-api-fluent-en_GB-jvm/src/test/kotlin/ch/tutteli/atrium/api/fluent/en_GB/PathFeatureExpectationsSpec.kt
@@ -1,38 +1,0 @@
-package ch.tutteli.atrium.api.fluent.en_GB
-
-import ch.tutteli.atrium.creating.Expect
-import ch.tutteli.atrium.specs.*
-import java.nio.file.Path
-
-class PathFeatureExpectationsSpec : ch.tutteli.atrium.specs.integration.PathFeatureExpectationsSpec(
-    property<Path, Path>(Expect<Path>::parent),
-    fun1<Path, Expect<Path>.() -> Unit>(Expect<Path>::parent),
-    feature1<Path, String, Path>(Expect<Path>::resolve),
-    fun2<Path, String,  Expect<Path>.() -> Unit>(Expect<Path>::resolve),
-    property<Path, String>(Expect<Path>::fileName),
-    fun1<Path, Expect<String>.() -> Unit>(Expect<Path>::fileName),
-    property<Path, String>(Expect<Path>::fileNameWithoutExtension),
-    fun1<Path, Expect<String>.() -> Unit>(Expect<Path>::fileNameWithoutExtension),
-    property<Path, String>(Expect<Path>::extension),
-    fun1<Path, Expect<String>.() -> Unit>(Expect<Path>::extension)
-) {
-    @Suppress("unused", "UNUSED_VALUE")
-    private fun ambiguityTest() {
-        var a1: Expect<DummyPath> = notImplemented()
-
-        a1.parent
-        a1 = a1.parent { }
-
-        a1.fileName
-        a1 = a1.fileName { }
-
-        a1.fileNameWithoutExtension
-        a1 = a1.fileNameWithoutExtension { }
-
-        a1.extension
-        a1 = a1.extension { }
-
-        a1.resolve("test")
-        a1.resolve("test", {})
-    }
-}
